+++ crates/client/rpc/src/versions/v0_7_1/methods/read/get_state_update.rs
@@ -61,7 +61,7 @@ pub fn get_state_update(starknet: &Starknet, block_id: BlockId) -> StarknetRpcRe
 
             // Get the old root from the previous block if it exists, otherwise default to zero.
             let old_root = if let Some(val) = block_info.header.block_number.checked_sub(1) {
-                let prev_block_info = &starknet.get_block_info(&DbBlockId::BlockN(val))?;
+                let prev_block_info = &starknet.get_block_info(&DbBlockId::Number(val))?;
                 let prev_block_info =
                     prev_block_info.as_nonpending().ok_or_internal_server_error("Block should not be pending")?;